+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-04-10 17:48:50
-// Commit: chore(ui): update dependencies and cleanup unused imports in ui
+// Updated at: 2026-04-10 18:16:22
+// Commit: feat(ui): add bi-directional event bus handler for ui
 
 class uiComponentState {
-  final String commitRef = '291';
-  final String lastMessage = 'chore(ui): update dependencies and cleanup unused imports in ui';
-  final DateTime timestamp = DateTime.parse('2026-04-10T17:48:50+05:00');
+  final String commitRef = '292';
+  final String lastMessage = 'feat(ui): add bi-directional event bus handler for ui';
+  final DateTime timestamp = DateTime.parse('2026-04-10T18:16:22+05:00');
 }